requirements.txt
alembic==1.12.0
bidict==0.22.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.0.0.post2
Flask==2.3.3
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
mysql-connector==2.2.9
python-engineio==4.7.1
python-socketio==5.9.0
requests==2.31.0
simple-websocket==0.10.1
SQLAlchemy==2.0.20
typing_extensions==4.7.1
urllib3==2.0.6
Werkzeug==2.3.7
wsproto==1.2.0
WTForms==3.0.1
alembic==1.12.0
annotated-types==0.6.0
anyio==3.7.1
bidict==0.22.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.0.0.post2
fastapi==0.104.0
Flask==2.3.3
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.18.0
httptools==0.6.1
httpx==0.25.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
mysql-connector==2.2.9
mysql-connector-python==8.1.0
orjson==3.9.9
protobuf==4.21.12
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
python-dotenv==1.0.0
python-engineio==4.7.1
python-multipart==0.0.6
python-socketio==5.9.0
PyYAML==6.0.1
requests==2.31.0
simple-websocket==0.10.1
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
typing_extensions==4.8.0
ujson==5.8.0
urllib3==2.0.6
uvicorn==0.23.2
watchfiles==0.21.0
websockets==12.0
Werkzeug==2.3.7
wsproto==1.2.0
WTForms==3.0.1
